[FIX] renaming testing utils
